VIITE-3037 AddressChangeType moved from RoadwayChangesDAO to its own file, under viite/model
