+++ viite-backend/base/src/main/scala/fi/vaylavirasto/viite/model/AddressChangeType.scala
@@ -1,0 +1,34 @@
+package fi.vaylavirasto.viite.model
+
+sealed trait AddressChangeType {
+  def value: Int
+}
+
+object AddressChangeType {
+  val values = Set(Unchanged, New, Transfer, ReNumeration, Termination)
+
+  def apply(intValue: Int): AddressChangeType = {
+    values.find(_.value == intValue).getOrElse(Unknown)
+  }
+
+  /*
+      Unchanged is a no-operation, tells TR that some road part or section stays intact but it needs
+        to be included in the message for other changes
+      New is a road address placing to a road that did not have road address before
+      Transfer is an adjustment of a road address, such as extending a road 100 meters from the start:
+        all the addresses on the first part are transferred with +100 to each start and end address M values.
+      ReNumeration is a change in road addressing but no physical or length changes. A road part gets a new
+        road and/or road part number.
+      Termination is for ending a road address (and possibly assigning the previously used road address
+        to a new physical location at the same time)
+   */
+
+  case object NotHandled extends AddressChangeType { def value = 0 }
+  case object Unchanged extends AddressChangeType { def value = 1 }
+  case object New extends AddressChangeType { def value = 2 }
+  case object Transfer extends AddressChangeType { def value = 3 }
+  case object ReNumeration extends AddressChangeType { def value = 4 }
+  case object Termination extends AddressChangeType { def value = 5 }
+  case object Unknown extends AddressChangeType { def value = 99 }
+
+}
