+++ viite-backend/viite-main/src/main/scala/fi/liikennevirasto/viite/RoadAddressService.scala
@@ -6,14 +6,13 @@ import fi.liikennevirasto.digiroad2.postgis.PostGISDatabase
 import fi.liikennevirasto.digiroad2.service.RoadLinkService
 import fi.liikennevirasto.digiroad2.util.LogUtils.time
 import fi.liikennevirasto.viite.dao.{RoadwayPointDAO, _}
-import fi.liikennevirasto.viite.dao.AddressChangeType.{ReNumeration, Termination, Transfer, Unchanged}
 import fi.liikennevirasto.viite.dao.CalibrationPointDAO.{CalibrationPointLocation, CalibrationPointType}
 import fi.liikennevirasto.viite.model.RoadAddressLink
 import fi.liikennevirasto.viite.process._
 import fi.liikennevirasto.viite.process.RoadAddressFiller.ChangeSet
 import fi.liikennevirasto.viite.util.CalibrationPointsUtils
 import fi.vaylavirasto.viite.geometry.{BoundingRectangle, GeometryUtils}
-import fi.vaylavirasto.viite.model.{BeforeAfter, Discontinuity, RoadLink, SideCode, Track}
+import fi.vaylavirasto.viite.model.{AddressChangeType, BeforeAfter, Discontinuity, RoadLink, SideCode, Track}
 import org.joda.time.DateTime
 import org.slf4j.LoggerFactory
 
@@ -957,7 +956,10 @@ class RoadAddressService(
 
     try {
 
-      val projectRoadwayChanges = roadwayChanges.filter(rw => List(Transfer, ReNumeration, Unchanged, Termination).contains(rw.changeInfo.changeType))
+      val projectRoadwayChanges = roadwayChanges.filter(rw => List(
+        AddressChangeType.Transfer, AddressChangeType.ReNumeration,
+        AddressChangeType.Unchanged, AddressChangeType.Termination
+      ).contains(rw.changeInfo.changeType))
       val updatedRoadwayPoints: Seq[RoadwayPoint] = projectRoadwayChanges.sortBy(_.changeInfo.target.startAddressM).foldLeft(
         Seq.empty[RoadwayPoint]) { (list, rwc) =>
         val change = rwc.changeInfo
@@ -968,7 +970,7 @@ class RoadAddressService(
             (source.startRoadPartNumber.get to source.endRoadPartNumber.get contains entry.roadPartNumber) &&
             entry.originalStartAddr >= source.startAddressM.get && entry.originalEndAddr <= source.endAddressM.get
         }
-        val roadwayNumbers = if (change.changeType == Termination) {
+        val roadwayNumbers = if (change.changeType == AddressChangeType.Termination) {
           terminatedRoadwayNumbersChanges.map(_.newRoadwayNumber).distinct
         } else {
           val roadwayNumbersInOriginalRoadPart = projectLinkChanges.filter(lc => lc.originalRoadNumber == source.roadNumber.get && lc.originalRoadPartNumber == source.startRoadPartNumber.get && lc.status.value == change.changeType.value)
@@ -986,7 +988,7 @@ class RoadAddressService(
         }.distinct
 
         if (roadwayPoints.nonEmpty) {
-          if (change.changeType == Transfer || change.changeType == Unchanged) {
+          if (change.changeType == AddressChangeType.Transfer || change.changeType == AddressChangeType.Unchanged) {
             if (!change.reversed) {
               val rwPoints: Seq[RoadwayPoint] = roadwayPoints.flatMap { rwp =>
 
@@ -1015,7 +1017,7 @@ class RoadAddressService(
               }
               list ++ rwPoints
             }
-          } else if (change.changeType == ReNumeration) {
+          } else if (change.changeType == AddressChangeType.ReNumeration) {
             if (change.reversed) {
               val rwPoints: Seq[RoadwayPoint] = roadwayPoints.flatMap { rwp =>
                 val newAddrM = Seq(source.endAddressM.get, target.endAddressM.get).max - rwp.addrMValue
@@ -1025,7 +1027,7 @@ class RoadAddressService(
             } else {
               list
             }
-          } else if (change.changeType == Termination) {
+          } else if (change.changeType == AddressChangeType.Termination) {
             val rwPoints: Seq[RoadwayPoint] = roadwayPoints.flatMap { rwp =>
               val terminatedRoadAddress = terminatedRoadwayNumbersChanges.find(change => change.originalRoadwayNumber == rwp.roadwayNumber &&
                 change.originalStartAddr >= source.startAddressM.get && change.originalEndAddr <= source.endAddressM.get
