+++ viite-backend/viite-main/src/main/scala/fi/liikennevirasto/viite/dao/RoadwayChangesDAO.scala
@@ -3,7 +3,7 @@ package fi.liikennevirasto.viite.dao
 import fi.liikennevirasto.digiroad2.dao.Sequences
 import fi.liikennevirasto.viite.process.{ProjectDeltaCalculator, RoadwaySection}
 import fi.liikennevirasto.viite.process.ProjectDeltaCalculator.{createTwoTrackOldAddressRoadParts, projectLinkDAO}
-import fi.vaylavirasto.viite.model.{AdministrativeClass, Discontinuity, LinkStatus, Track}
+import fi.vaylavirasto.viite.model.{AddressChangeType, AdministrativeClass, Discontinuity, LinkStatus, Track}
 
 import java.sql.{PreparedStatement, Timestamp}
 import org.joda.time.DateTime
@@ -12,40 +12,6 @@ import org.slf4j.LoggerFactory
 import slick.driver.JdbcDriver.backend.Database.dynamicSession
 import slick.jdbc.{GetResult, PositionedResult, StaticQuery => Q}
 import slick.jdbc.StaticQuery.interpolation
-
-
-sealed trait AddressChangeType {
-  def value: Int
-}
-
-object AddressChangeType {
-  val values = Set(Unchanged, New, Transfer, ReNumeration, Termination)
-
-  def apply(intValue: Int): AddressChangeType = {
-    values.find(_.value == intValue).getOrElse(Unknown)
-  }
-
-  /*
-      Unchanged is a no-operation, tells TR that some road part or section stays intact but it needs
-        to be included in the message for other changes
-      New is a road address placing to a road that did not have road address before
-      Transfer is an adjustment of a road address, such as extending a road 100 meters from the start:
-        all the addresses on the first part are transferred with +100 to each start and end address M values.
-      ReNumeration is a change in road addressing but no physical or length changes. A road part gets a new
-        road and/or road part number.
-      Termination is for ending a road address (and possibly assigning the previously used road address
-        to a new physical location at the same time)
-   */
-
-  case object NotHandled extends AddressChangeType { def value = 0 }
-  case object Unchanged extends AddressChangeType { def value = 1 }
-  case object New extends AddressChangeType { def value = 2 }
-  case object Transfer extends AddressChangeType { def value = 3 }
-  case object ReNumeration extends AddressChangeType { def value = 4 }
-  case object Termination extends AddressChangeType { def value = 5 }
-  case object Unknown extends AddressChangeType { def value = 99 }
-
-}
 
 case class RoadwayChangeSection(roadNumber: Option[Long], trackCode: Option[Long], startRoadPartNumber: Option[Long],
                                 endRoadPartNumber: Option[Long], startAddressM: Option[Long], endAddressM: Option[Long],
